+++ src/Grid/BlockingPoint.java
@@ -2,7 +2,7 @@ package Grid;
 
 public class BlockingPoint extends GridPoint 
 {
-	public BlockingPoint(int XCoord, int YCoord)
+	public BlockingPoint(float XCoord, float YCoord)
 	{
 		this.xCoord = XCoord;
 		this.yCoord = YCoord;
